Added alerts endpoint, reorganized files
